requirements.txt
fastapi==0.73.0
numpy==1.20.3
pandas==1.5.3
uvicorn==0.15.0
image==1.5.33
scikit-learn==1.2.2
python-dotenv==1.0.0